refactor: fix UsersRepository import paths in deactivate-user-use-case and prisma-users-repository
