+++ src/core/membership/use-cases/deactivate-user-use-case.ts
@@ -1,5 +1,6 @@
 import { Id } from '@/core/global/domain/structures'
-import { UseCase, UsersRepository } from '@/core/global/interfaces'
+import { UseCase } from '@/core/global/interfaces'
+import { UsersRepository } from '@/core/membership/interfaces'
 import { UserDto } from '../domain/entities/dtos'
 import { User } from '../domain/entities'
 import { UserNotFoundError, UserAlreadyDeactivatedError } from '../domain/errors'
@@ -20,8 +21,6 @@ export class DeactivateUserUseCase implements UseCase<Request, Response> {
     }
 
     user.deactivate()
-
-    console.log('user', user)
 
     await this.repository.replace(user)
     return user.dto
